refactor(dg): create layoutStore
