+++ packages/data-grid/src/stores/index.ts
@@ -2,3 +2,4 @@ export * from './data'
 export * from './pagination'
 export * from './filter'
 export * from './sort'
+export * from './layout'
